refactored truncation selection tests
